Remove pycache
